Add normalizeRoute tests, change the regex match to test for label matching
