+++ public/app/features/alerting/unified/routeGroupsMatcher.worker.ts
@@ -2,11 +2,11 @@ import * as comlink from 'comlink';
 
 import type { AlertmanagerGroup, RouteWithID } from '../../../plugins/datasource/alertmanager/types';
 
-import { findMatchingAlertGroups, normalizeRootRoute } from './utils/notification-policies';
+import { findMatchingAlertGroups, normalizeRoute } from './utils/notification-policies';
 
 const routeGroupsMatcher = {
   getRouteGroupsMap(rootRoute: RouteWithID, groups: AlertmanagerGroup[]): Map<string, AlertmanagerGroup[]> {
-    const normalizedRootRoute = normalizeRootRoute(rootRoute);
+    const normalizedRootRoute = normalizeRoute(rootRoute);
 
     function addRouteGroups(route: RouteWithID, acc: Map<string, AlertmanagerGroup[]>) {
       const routeGroups = findMatchingAlertGroups(normalizedRootRoute, route, groups);
